small refactors
